+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### 💡 Others
 
+- Add CLI archive build phase to macOS project. ([#248](https://github.com/expo/orbit/pull/248) by [@gabrieldonadel](https://github.com/gabrieldonadel))
+
 ## 2.0.2 — 2025-02-27
 
 ### 🎉 New features